+++ raganything/__init__.py
@@ -1,7 +1,7 @@
 from .raganything import RAGAnything as RAGAnything
-from .raganything import RAGAnythingConfig as RAGAnythingConfig
+from .config import RAGAnythingConfig as RAGAnythingConfig
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Zirui Guo"
 __url__ = "https://github.com/HKUDS/RAG-Anything"
